Threading improvements
Used PyQt5's built in threading to hand the issue of locking the GUI when the the water flow sensor was collecting data, This was caused by the loop running infinitely in the main thread as the GUI. By using threading we were able to run the flow sensor on a separate thread and update the GUI with what the flow meter was recording.
A few UI changes occurred with positioning the start and stop buttons to be almost exact when switching pages from OFF to ON mode. Also allowed the current to be changed automatically through the threading and sized it properly.